Most logic in hosts now shared.
