+++ src/PerformanceTests/NServiceBus5/App.config
@@ -9,6 +9,8 @@
   <appSettings>
     <add key="SplunkURL" value="deploy.particular.net" />
     <add key="SplunkPort" value="8001"/>
+    <add key="WarmupDuration" value="00:00:03"/>
+    <add key="RunDuration" value="00:00:10"/>
   </appSettings>
 
   <MessageForwardingInCaseOfFaultConfig ErrorQueue="error"/>
