+++ src/PerformanceTests/NServiceBus5/NServiceBus5.csproj
@@ -7,7 +7,7 @@
     <ProjectGuid>{BE2C7B5E-7A71-4B0C-9020-343ED2010993}</ProjectGuid>
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>NServiceBus5</RootNamespace>
+    <RootNamespace>Host</RootNamespace>
     <AssemblyName>NServiceBus5</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
@@ -33,7 +33,7 @@
     <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <PropertyGroup>
-    <StartupObject>NServiceBus5.Program</StartupObject>
+    <StartupObject>Host.Program</StartupObject>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="log4net, Version=1.2.10.0, Culture=neutral, PublicKeyToken=1b44e1d426115821, processorArchitecture=MSIL">
@@ -104,15 +104,14 @@
   <ItemGroup>
     <Compile Include="ConcurrencyProfile.cs" />
     <Compile Include="IProfile.cs" />
-    <Compile Include="Program.cs" />
+    <Compile Include="Program.V5.cs" />
     <Compile Include="TransactionProfile.cs" />
   </ItemGroup>
   <Import Project="..\Common.Data\Common.Data.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
-    <PostBuildEvent>copy "$(TargetPath)" "$(TargetDir)NServiceBus5.x86.exe"
-copy "$(TargetPath).config" "$(TargetDir)NServiceBus5.x86.exe.config"
-"%25programFiles(x86)%25\Microsoft SDKs\Windows\v10.0A\bin\NETFX 4.6.1 Tools\CorFlags.exe" "$(TargetDir)NServiceBus5.x86.exe" /32bit+</PostBuildEvent>
+    <PostBuildEvent>
+    </PostBuildEvent>
   </PropertyGroup>
   <PropertyGroup Label="SlowCheetah">
     <SlowCheetahToolsPath>$([System.IO.Path]::GetFullPath( $(MSBuildProjectDirectory)\..\..\..\packages\SlowCheetah.2.5.15\tools\))</SlowCheetahToolsPath>
